+++ CHANGELOG.md
@@ -1,5 +1,7 @@
+
 ### Unreleased
 
+- Imported resolvers will delegate to the imported component schema to which they belong.
 - Remove `this._context` as default value for context in `execute()` requiring `execute()` users to pass in context from a calling resolver.
 
 ### v1.2.4